Closes #9 - User can enter newline where they want
Before, the enter key was always hijacked and a new line would always be
added below the comments, regardless of where on the line they were
hitting enter.
This fixes things so that they can hit enter wherever they want and it
will act like normal.
If they hit enter at the end of a line, the new line will be added after
the following comments, to make sure the comments stay with the context
they are meant to have.