Hide all comment icons when inserting new data
HTML might get sent to the server that includes one of the comment icons
being visible. When this is sent to the other users, they'll see this
icon even though they aren't hovering over it. This clears any shown
icons when the page data is put into the user's document.